+++ client/rest-high-level/src/main/java/org/elasticsearch/client/ml/GetBucketsRequest.java
@@ -86,7 +86,7 @@ public class GetBucketsRequest extends ActionRequest implements ToXContentObject
 
     /**
      * Sets the timestamp of a specific bucket to be retrieved.
-     * @param timestamp the timestamp of a specific bucket to be retrieved
+     * @param timestamp String representation of a timestamp; may be an epoch seconds, epoch millis or an ISO string
      */
     public void setTimestamp(String timestamp) {
         this.timestamp = timestamp;
@@ -129,7 +129,7 @@ public class GetBucketsRequest extends ActionRequest implements ToXContentObject
     /**
      * Sets the value of "start" which is a timestamp.
      * Only buckets whose timestamp is on or after the "start" value will be returned.
-     * @param start value of "start" to be set
+     * @param start String representation of a timestamp; may be an epoch seconds, epoch millis or an ISO string
      */
     public void setStart(String start) {
         this.start = start;
@@ -142,7 +142,7 @@ public class GetBucketsRequest extends ActionRequest implements ToXContentObject
     /**
      * Sets the value of "end" which is a timestamp.
      * Only buckets whose timestamp is before the "end" value will be returned.
-     * @param end value of "end" to be set
+     * @param end String representation of a timestamp; may be an epoch seconds, epoch millis or an ISO string
      */
     public void setEnd(String end) {
         this.end = end;
